lsbstego/requirements.txt
pillow==11.3.0
pycryptodome==3.23.0
